backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
python-dotenv==1.0.0
python-multipart==0.0.6
google-generativeai==0.3.2
pymongo==4.6.0
PyMuPDF==1.23.10
pdfplumber==0.10.3
pdf2image==1.16.3
camelot-py[cv]==0.10.1
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.1
opencv-python==4.8.1.78
pytesseract==0.3.10
scikit-learn==1.3.2
layoutparser==0.3.4
easyocr==1.7.0
spacy==3.7.2
nltk==3.8.1
requests==2.31.0
aiofiles==23.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2